Create REAR LIGHTS OF CAR USING ARDUINO
This task was indeed an interesting one. I learned how to build a rear light which is used nowadays in most of the cars. It was quite difficult at starting, when I had to use the pushbutton and the blinking state at the same time. But after working hard and coding I came to the point where I was successful with the task.