Create Building a twilight switch using LDR
In this task I became familiar with photoresistors. It is used mostly to indicate the presence or absence of light or to measure light intensity. I connected the led and photoresistors to the breadboard with 2 resistors (10k and 220k ohm) and Arduino. It was overall an easy and interesting task.