Create Adjusting the brightness of the led using a potentiometer.
In this task, I became familiar with potentiometer. It is a great tool to adjust the output voltage so the sliding contact gets moved along the resistor on the output side. I connected the led, the potentiometer, and a resistor with wires to the breadboard and Arduino. I got in trouble while connecting the legs of the pot to Arduino. Then I researched further about the legs and how it works. Thus, I became successful in completing the task.